tests/requirements.txt
# Testing dependencies for the text annotation system

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0  # Parallel test execution

# Testing utilities
httpx==0.25.2  # Async HTTP client for testing
requests==2.31.0
responses==0.24.1  # HTTP request mocking

# Database testing
sqlalchemy-utils==0.41.1
alembic==1.12.1

# File and data testing
Pillow==10.1.0  # Image processing for testing
openpyxl==3.1.2  # Excel file testing
pandas==2.1.4

# Mock and fixture utilities
factory-boy==3.3.0
faker==20.1.0
freezegun==1.2.2  # Time/date mocking

# Performance testing
pytest-benchmark==4.0.0
memory-profiler==0.61.0

# Reporting
pytest-html==4.1.1  # HTML test reports
pytest-json-report==1.5.0  # JSON test reports